// packages/rpc/src/types/api.d.ts
 declare namespace CKBComponents {
// ...
  export interface Node {
    url: string
    httpAgent?: any
    httpsAgent?: any
  }
  export interface Method {
    name: string
    method: string
    paramsFormatters: Function[]
    resultFormatters?: Function
  }
// ...
}